# fix(db): add index to configs