--- a/packages/shared/lib/db/migrations/20240222182357_configs-index.cjs
+++ b/packages/shared/lib/db/migrations/20240222182357_configs-index.cjs
@@ -1,0 +1,11 @@
+exports.config = { transaction: false };
+
+exports.up = function (knex) {
+    return knex.schema.raw(
+        'CREATE INDEX "idx_configs_environmentid_uniquekey_deleted" ON "nango"."_nango_configs" USING BTREE ("environment_id","unique_key") WHERE deleted = false'
+    );
+};
+
+exports.down = function (knex) {
+    return knex.schema.raw('DROP INDEX CONCURRENTLY idx_configs_environmentid_uniquekey_deleted');
+};